release 3.1.128 source code

+++ Services/Rds/V3/Model/Auditlog.cs
@@ -32,7 +32,7 @@ namespace HuaweiCloud.SDK.Rds.V3.Model
         /// 审计日志大小，单位：KB。
         /// </summary>
         [JsonProperty("size", NullValueHandling = NullValueHandling.Ignore)]
-        public long? Size { get; set; }
+        public double? Size { get; set; }
 
         /// <summary>
         /// 审计日志开始时间，格式为“yyyy-mm-ddThh:mm:ssZ”。  其中，T指某个时间的开始，Z指时区偏移量，例如北京时间偏移显示为+0800。
